Add info about the -Ss flag to README. Make sync more user friendly.
The README file now contains information about the -Ss flag (search for
package in gists). The -f (force changes) flag, that calls the sync
function, is now more user friendly: it lists packags that are going to
be removed and the new ones that going to be installed (if any), and
prints a message if there no packages to remove or no new packages to
install.