Fix bug related to file overwritting.
Files were being overwritten when syncronizing the packages between 2 or more machines: the temporary file used by the force changes function had the same name as the temp file used by the hook. This meant the file was being deleted and it wasn't possible to install new packages if there were also packages to delete.